requirements.txt
python-telegram-bot[all]==21.6
httpx==0.27.0
